test: operations tests (spec => test to match Deno's pattern)
